+++ jupyterlab_bigquery/src/components/list_items_panel/list_tree_item.tsx
@@ -109,7 +109,7 @@ export interface ProjectProps extends ResourceProps {
   openSnackbar: any;
   removeProject: any;
   collapseAll?: boolean;
-  updateCollapseAll? : any;
+  updateCollapseAll?: any;
 }
 
 interface ResourceState {
